+++ packages/blaze-core/src/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./util";
 export * from "./params";
 export * from "./pretty";
+export * from "./errors";