+++ packages/blaze-emulator/jest.config.ts
@@ -1,11 +1,2 @@
 import config from "@blaze-cardano/jest-config/base.jest.config";
-
-export default {
-  ...config,
-  transform: {
-    "^.+\\.(t|j)sx?$": ["ts-jest", { isolatedModules: true }],
-  },
-  transformIgnorePatterns: [
-    `<rootDir>/node_modules/.pnpm/(?!(@utxorpc\\+sdk)@)`,
-  ],
-};
+export default config;